requirements.txt
altgraph==0.17
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
av==8.0.2
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cmapy==0.6.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.1.0
importlib-resources==3.3.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
json-tricks==3.15.5
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
macholib==1.14
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.5
oauth2client==4.1.3
opencv-contrib-python==4.5.1.48
opencv-contrib-python-headless==4.4.0.46
opencv-python==4.5.1.48
opencv-python-headless==4.4.0.46
OpenVisus==2.1.106
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
PIMS==0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyExifTool==0.1.1
pygame==2.0.0
Pygments==2.7.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pymap3d==2.4.3
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySimpleGUI==4.32.1
pysolar==0.9
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzbar==0.1.8
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
slampy==1.0.1
slicerator==1.0.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
tk==0.1.0
tornado==6.1
trackpy==0.4.2
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
